automated-learning-system/backend/requirements.txt
fastapi==0.100.0
uvicorn[standard]==0.22.0
SQLAlchemy==2.0.22
alembic==1.11.1
pydantic==1.10.12
pydantic[email]
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
websockets==11.0.3
gymnasium==0.28.1
stable-baselines3==2.0.0
torch==2.2.0
torchvision==0.17.0
transformers==4.35.2
numpy==1.26.2
pandas==2.1.2
aiofiles==23.1.0
python-dotenv==1.0.0
typing_extensions==4.8.0
email-validator==1.3.1
requests>=2.28
librosa
opencv-python
deepface
Pillow
httpx>=0.23.0,<0.28.0
pytest
onnxruntime
# Dataset download utility (optional, needed for download_datasets.py)
# kaggle
